Allow PippoFilter to auto-commit uncommitted responses
All responses start out as 200 OK from the underlying container because they have
been matched by PippoFilter.  A Pippo application must then handle that response and
either preserve the 200 or set it differently.

If the handler chain does not explicitly commit the response, PippoFilter should treat
this as a successful operation, like the underlying servlet framework would.

In the event that the uncommitted response has an error status (>= 400) then the
error handler should be called to return the appropriate representation of that error:
xml, json, html, etc.

The auto-commit behavior should be logged and the final status code of all responses
should be logged too.