Fix "response already committed" bug in file and resource handling
When we grab the InputStream of the response we have implicitly
committed the response.  We were then trying to set the contentLength
header which caused the error.  Setting that header is pointless anyway
because the next step is to flush the buffer which changes the response
to chunked encoding which discards contentLength headers anyway.

The solution is to not set the contentLength header.