Fix another Response already committed bug
When we grab the output stream/writer or flush the buffer we have initiated committing
the Response.  That means we should not re-check isCommitted() within the sequence else we throw an exception.

In this case we were "appending" the generated template to the Response writer and then committing cookies, etc afterwards.

This would fail due to the following sequence:

1. send(CharSequence)
2. checkCommitted()
3. append(CharSequence)
4. checkCommitted()
5. response.getWriter().append(sequence)
6. commit()
7. checkCommitted() -> boom